[dev] convert from Utils to utils and rename folder from nodeMailer to node-mailer
